Revert "s/ReadLong/ReadInt/ to match the return type"
This reverts commit 3d771a6882f4cea7790ce8e3467bf33d0e2e607b.

+++ neo/d3xp/Fx.cpp
@@ -899,8 +899,8 @@ void idEntityFx::ReadFromSnapshot( const idBitMsg& msg )
 	
 	GetPhysics()->ReadFromSnapshot( msg );
 	ReadBindFromSnapshot( msg );
-	fx_index = gameLocal.ClientRemapDecl( DECL_FX, msg.ReadInt() );
-	start_time = msg.ReadInt();
+	fx_index = gameLocal.ClientRemapDecl( DECL_FX, msg.ReadLong() );
+	start_time = msg.ReadLong();
 	
 	if( fx_index != -1 && start_time > 0 && !fxEffect && started < 0 )
 	{
